Create BST Tree, Create AVL Tree

+++ src/B_Cliente/objects/Player2.java
@@ -1,19 +1,22 @@
 package B_Cliente.objects;
 
+import B_Cliente.main.Game;
+
 import java.awt.*;
+import java.awt.event.MouseEvent;
+import java.awt.event.MouseListener;
+import java.awt.event.MouseMotionListener;
 import java.awt.image.BufferedImage;
 import java.util.LinkedList;
 
-public class Player2 extends GameObject{
-
-    public Player2(float x, float y, ObjectId Id) {
+public class Player2 extends GameObject {
+    public Player2(int x, int y, ObjectId Id) {
         super(x, y, Id);
     }
 
     @Override
     public void tick(LinkedList<GameObject> object) {
     }
-
     @Override
     public void render(Graphics g) {
         BufferedImage playerImg = Load.imageLoad("/player.png");
@@ -64,4 +67,5 @@ public class Player2 extends GameObject{
     public ObjectId getId() {
         return null;
     }
+
 }
